requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.3
django-cors-headers==4.3.1
sqlparse==0.5.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0
